+++ NEWS.md
@@ -1,4 +1,11 @@
 # ICAMS 2.1.0.900x
+* Slight changes to behavior of PlotCatalog() and PlotCatalogToPdf() for class
+  SBS96Catalog: (New) Allow setting ylim and cex.
+  (New) For PlotCatalog (not PlotCatalogToPdf), allow plotting of a 96 x 2 catalog,
+  in which case behavior is a stacked
+  bar chart. (New) Plot x axis tick marks if xlabels is not TRUE; set
+  \code{par(tck = 0)} to suppress. 
+  ? Minor changes in default font sizes.
 * (Non backward-compatible.) Changed the return of internal functions
   ReadMutectVCF(), ReadStrelkaSBSVCF() and ReadStrelkaIDVCF() to a list of two
   objects. The first object is a data frame storing data lines of the VCF file.
@@ -10,8 +17,7 @@
   number of rows in the first object.
 * (Non backward-compatible.) Changed the return of exported functions
   ReadAndSplitMutectVCFs(), ReadAndSplitStrelkaSBSVCFs() to a list of two lists. 
-* (Non backward-compatible.)
-  Added a new dependency package "zip" in ICAMS, to be used in three new exported  
+* Added a new dependency package "zip" in ICAMS, to be used in three new exported  
   functions MutectVCFFilesToZipFile(), StrelkaSBSVCFFilesToZipFile() and 
   StrelkaIDVCFFilesToZipFile(). 
 * Renamed internal function creating custom abundance to GetCustomKmerCounts().